api/requirements-3.10.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.1
Markdown==3.6
psycopg2==2.9.9
python-environ==0.4.54
sqlparse==0.4.4
typing_extensions==4.10.0
